pyproject.toml
[project]
name = "rblxapi2cpp"
version = "0.1.0"
description = "Tool to convert Roblox api docs to C++ classes"
authors = [
    {name = "RubberSpring",email = "143825964+RubberSpring@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry.scripts]
rblxapi2cpp = "rblxapi2cpp.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
